requirements.txt
discord
python-dotenv
flask